make dashboard change modified:   ecommerce-app/src/components/elements/UserTable. modified:   ecommerce-app/src/components/fragments/dashboard/Navbar. modified:   ecommerce-app/src/features/product/useDeleteProduct. modified:   ecommerce-app/src/pages/admin/dashboard/Dashboard. deleted:    ecommerce-app/src/pages/admin/dashboard/DashboardTest. modified:   ecommerce-app/src/router.jsx
