+++ ecommerce-app/src/components/fragments/dashboard/Navbar.jsx
@@ -30,12 +30,6 @@ export default function Navbar() {
                 <div className="flex justify-start items-center w-auto mr-10">
                     <NavLink text={"Dashboard"} href={"/dashboard"} />
                 </div>
-                <div className="flex justify-start items-center w-auto mr-10">
-                    <NavLink text={"Product List"} href={"/dashboard/product"} />
-                </div>
-                <div className="flex justify-start items-center w-auto mr-10">
-                    <NavLink text={"Dashboard Test"} href={"/dashboard/test"} />
-                </div>
             </div>
