+++ ecommerce-app/src/router.jsx
@@ -11,7 +11,6 @@ import Register from "./pages/auth/Register";
 import ShoppingCart from "./pages/guest/ShoppingCart"
 import DashboardProductDetail from "./pages/admin/dashboard/DashboardProductDetail";
 import DashboardProductEdit from "./pages/admin/dashboard/DashboardProductEdit";
-import DashboardTest from "./pages/admin/dashboard/DashboardTest";
 
 export const router = createBrowserRouter([
     {
@@ -39,10 +38,6 @@ export const router = createBrowserRouter([
             {
                 path: "product",
                 element: <DashboardProduct />,
-            },
-            {
-                path: "test",
-                element: < DashboardTest/>,
             },
             {
                 path: "create",
